--- BoxOffice/Dto/BoxOfficeResultDTO.swift
@@ -1,11 +1,0 @@
-import Foundation
-
-struct BoxOfficeResultDTO: Decodable {
-    private let boxOfficeResult: BoxOfficeDTO
-}
-
-extension BoxOfficeResultDTO {
-    func boxOfficeDTO() -> BoxOfficeDTO {
-        return boxOfficeResult
-    }
-}
